Revamp outro animation
Use the wallpaper as part of the window background rather than trying
to replicate it with WallpaperManager. This allows us to perform a
circular hide to whatever the wallpaper currently looks like exactly,
rather than needing to obtain the wallpaper Bitmap or Drawable and
guess. Display the Finish activity edge-to-edge. Fade out rather
than slide out.

Change-Id: I1154043178a04ca0ced7e8365e338dc10f3377cb
Signed-off-by: Aditya Eka Pranata <aepranata@gmail.com>